Fix stale countdowns, harden security, and split page into components
- Untrack .env.local, ignore .env*, drop unused cheerio dependency
- Fix critical/high npm advisories (vitest 3.2.7), add @types/luxon
- Enable TypeScript strict mode
- Refresh countdowns every 30s so relative times stop freezing
- Sync <html lang> with the selected language; apply theme before paint
- Decompose app/page.tsx (1146 lines) into components and hooks
- Split lib/language.ts (1246 lines) into per-language modules
- Remove dead i18n keys, exports, and identity track aliases
- Handle RFC 5545 folded lines in the ICS parser (+tests)
- CI: least-privilege deploy permissions, typecheck+test steps, PR workflow
