--- .env.local
@@ -1,14 +1,0 @@
-# Firebase web app configuration
-NEXT_PUBLIC_FIREBASE_API_KEY=AIzaSyAcG6THP3WIIYBSgcgISA2MCbt-7K96Z_U
-NEXT_PUBLIC_FIREBASE_AUTH_DOMAIN=racesync-a454b.firebaseapp.com
-NEXT_PUBLIC_FIREBASE_PROJECT_ID=racesync-a454b
-NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET=racesync-a454b.firebasestorage.app
-NEXT_PUBLIC_FIREBASE_MESSAGING_SENDER_ID=969154671211
-NEXT_PUBLIC_FIREBASE_APP_ID=1:969154671211:web:f9c22f65b2f8d662b7f72e
-NEXT_PUBLIC_FIREBASE_MEASUREMENT_ID=G-4FSRBEZBXS
-
-# Firebase Cloud Messaging VAPID key
-NEXT_PUBLIC_FIREBASE_VAPID_KEY=BOmFcJ8LuPBjkIh9Erljbpx7-PQk8pH6fyKwsfrOBAJHTalPl_-JgpGBVUaYK__D_vkQlOyS2PAc3NQHD0UTiRA
-
-# Optional: Firestore collection for storing messaging tokens
-NEXT_PUBLIC_FIREBASE_TOKENS_COLLECTION=pushTokens
