+++ .env.local.example
@@ -1,2 +1,10 @@
-# Нет обязательных переменных окружения для базовой работы сайта.
-# При необходимости добавляйте свои NEXT_PUBLIC_* переменные здесь.
+# Переменные окружения проекта.
+# Все NEXT_PUBLIC_* переменные попадают в клиентский бандл — секретов здесь быть не должно.
+
+# Заполняется автоматически в GitHub Actions; локально можно не задавать.
+# Используется next.config.js для вычисления basePath/assetPrefix при деплое на GitHub Pages.
+# GITHUB_REPOSITORY=owner/repo
+# GITHUB_REPOSITORY_OWNER=owner
+
+# Опциональный префикс для статических ассетов (обычно совпадает с basePath).
+# NEXT_PUBLIC_ASSET_PREFIX=/my-repo
