+++ lib/language/ru.ts
@@ -1,0 +1,178 @@
+import type { LanguageDefinition } from './types';
+import { formatCountdownStart } from './countdown';
+
+export const ru: LanguageDefinition = {
+  code: 'ru',
+  name: 'Русский',
+  shortName: 'RU',
+  locale: 'ru',
+  periodOptions: [
+    { label: '24 часа', value: 24 },
+    { label: '48 часов', value: 48 },
+    { label: '72 часа', value: 72 },
+    { label: '7 дней', value: 168 },
+    { label: '30 дней' },
+  ],
+  sessionLabels: {
+    Qualifying: 'Квалификация',
+    Race: 'Гонка',
+    Sprint: 'Спринт',
+  },
+  texts: {
+    heroTitle: 'My race weekend',
+    heroSubtitle:
+      'Синхронизируйтесь с динамикой гоночных уик-эндов: фильтруйте серии, управляйте горизонтом просмотра и следите за временем старта в собственном часовом поясе.',
+    seriesLabel: 'Серии',
+    activeSelection: names => `Выбрано: ${names.join(' · ')}`,
+    allSeriesHidden: 'Все серии скрыты',
+    reviewPeriodLabel: 'Период обзора',
+    eventsInWindowLabel: 'Событий в окне',
+    nextStartLabel: 'Ближайший старт',
+    noEvents: 'Нет событий',
+    extendPeriodHint: 'Попробуйте расширить период',
+    countdownStart: relative => formatCountdownStart(relative, 'Старт', 'Старт через'),
+    countdownLive: relative => (relative ? `Идёт • старт ${relative}` : 'Идёт'),
+    countdownFinish: relative => `Финиш ${relative}`,
+    countdownScheduled: 'По расписанию',
+    trackLayoutLabel: parts =>
+      parts.length ? `Схема автодрома: ${parts.join(' — ')}` : 'Схема автодрома',
+    theme: {
+      toggleToDark: 'Переключить на тёмную тему',
+      toggleToLight: 'Переключить на светлую тему',
+    },
+    upcomingEventDescriptorFallback: 'Нет событий',
+    brandName: 'RaceSync',
+    navFeatures: 'Возможности',
+    navFaq: 'Вопросы',
+    heroCta: 'К событиям',
+    scheduleTitle: 'Лента уик-эндов',
+    scheduleSubtitle: 'Онлайн-обновление стартов с учётом вашего часового пояса.',
+    scheduleLoadingLabel: 'Загружаем расписание…',
+    scheduleErrorTitle: 'Не удалось загрузить расписание',
+    scheduleErrorDescription: 'Попробуйте ещё раз. Если ошибка повторится, откройте файл расписания вручную.',
+    scheduleRetryButton: 'Повторить',
+    scheduleErrorFallbackPrefix: 'Файл для диагностики:',
+    scheduleIcsLinkLabel: 'Открыть schedule.ics',
+    featuresTitle: 'Сила оперативного календаря',
+    featuresSubtitle: 'Всё, чтобы не пропустить старт.',
+    features: [
+      {
+        title: 'Локальное время без вычислений',
+        description: 'Все сессии автоматически отображаются в вашем часовом поясе, без ручного пересчёта.',
+      },
+      {
+        title: 'Гибкая фильтрация серий',
+        description: 'Выберите только интересующие чемпионаты и сосредоточьтесь на нужных событиях.',
+      },
+      {
+        title: 'Схемы трасс с деталями',
+        description: 'Каждый этап сопровождается контуром автодрома и кратким контекстом.',
+      },
+    ],
+    insightsTitle: 'Как это работает',
+    insightsSubtitle: 'Три шага до полной картины гоночного уик-энда.',
+    insightsSteps: [
+      {
+        title: 'Выберите интересующие серии',
+        description: 'Оставьте F1, F2, F3 или MotoGP — всё под контролем в один клик.',
+      },
+      {
+        title: 'Настройте период обзора',
+        description: 'Расширьте окно до 30 дней или сфокусируйтесь на ближайших 24 часах.',
+      },
+      {
+        title: 'Следите за обратным отсчётом',
+        description: 'Живые подсказки покажут, сколько осталось до старта или финиша.',
+      },
+    ],
+    faqTitle: 'Частые вопросы',
+    faqSubtitle: 'Быстрые ответы на популярные запросы сообщества.',
+    faqItems: [
+      {
+        question: 'Откуда берутся данные?',
+        answer:
+          'Мы синхронизируемся с официальными календарями чемпионатов и обновляем расписание автоматически.',
+      },
+      {
+        question: 'Можно ли пользоваться на телефоне?',
+        answer: 'Да, интерфейс адаптирован под мобильные устройства и сохраняет фильтры.',
+      },
+      {
+        question: 'Не вижу нужное событие — что делать?',
+        answer: 'Убедитесь, что серия включена, и расширьте период обзора до 30 дней.',
+      },
+    ],
+    ctaTitle: 'Готовы к старту?',
+    ctaSubtitle: 'Откройте живой календарь и держите расписание всегда под рукой.',
+    ctaButton: 'Открыть календарь',
+    footer: {
+      tagline: 'RaceSync помогает болельщикам синхронизироваться с гоночными уик-эндами.',
+      productHeading: 'Продукт',
+      resourcesHeading: 'Ресурсы',
+      supportHeading: 'Поддержка',
+      contactEmailLabel: 'Почта команды',
+      contactEmail: 'hello@racesync.app',
+      legal: '© {year} RaceSync. Все права защищены.',
+      productLinks: [
+        { label: 'Расписание', href: '#schedule' },
+        { label: 'Возможности', href: '#features' },
+        { label: 'Как это работает', href: '#insights' },
+      ],
+      resourcesLinks: [
+        { label: 'Частые вопросы', href: '#faq' },
+        { label: 'Календарь .ics', href: './schedule.ics' },
+      ],
+      supportLinks: [
+        { label: 'Написать нам', href: 'mailto:hello@racesync.app' },
+        { label: 'Политика конфиденциальности', href: '#privacy' },
+      ],
+    },
+    privacyPolicy: {
+      title: 'Политика конфиденциальности',
+      lastUpdated: 'Обновлено: 20 марта 2024 г.',
+      intro: [
+        'RaceSync создан, чтобы помогать болельщикам следить за расписанием гонок и при этом собирать минимум данных. В этом документе описано, какую информацию мы обрабатываем, когда вы пользуетесь сайтом или календарями.',
+      ],
+      sections: [
+        {
+          title: 'Какие данные мы собираем',
+          paragraphs: [
+            'Мы стремимся получать только те сведения, которые необходимы для работы сервиса.',
+          ],
+          list: [
+            'Контактные данные, которые вы добровольно отправляете на hello@racesync.app.',
+            'Обезличенную аналитику о посещениях страниц и том, какие функции используются.',
+            'Технические журналы, автоматически формируемые для защиты сайта от сбоев и злоупотреблений.',
+          ],
+        },
+        {
+          title: 'Как мы используем данные',
+          paragraphs: [
+            'Полученная информация помогает улучшать RaceSync и поддерживать стабильную работу сервиса.',
+          ],
+          list: [
+            'Отвечать на обращения и вопросы, которые вы отправляете по электронной почте.',
+            'Понимать, какие разделы и возможности востребованы у пользователей.',
+            'Выявлять технические проблемы и обеспечивать безопасность инфраструктуры.',
+          ],
+        },
+        {
+          title: 'Хранение и безопасность',
+          paragraphs: [
+            'Доступ к данным ограничен небольшой командой RaceSync. Аналитика хранится в агрегированном виде и не позволяет идентифицировать отдельных пользователей.',
+            'Контактные сообщения хранятся столько, сколько нужно для решения вашего запроса, после чего удаляются из активных систем.',
+          ],
+        },
+        {
+          title: 'Ваш выбор',
+          paragraphs: [
+            'Вы можете ограничить сбор обезличенной аналитики с помощью настроек браузера или расширений блокировки. Если хотите удалить ранее переданные данные, напишите нам — мы обработаем запрос в разумный срок.',
+          ],
+        },
+      ],
+      conclusion:
+        'Если у вас остались вопросы о защите данных в RaceSync, напишите на hello@racesync.app — мы ответим в кратчайшие сроки.',
+      closeLabel: 'Закрыть',
+    },
+  },
+};
