+++ lib/preferences.ts
@@ -1,3 +1,6 @@
 export const LANGUAGE_STORAGE_KEY = 'schedule-language';
 export const SERIES_STORAGE_KEY = 'schedule-visible-series';
 export const PERIOD_STORAGE_KEY = 'schedule-review-period-hours';
+
+/** Hour values exposed by the viewing-window controls ("30 days" is encoded as "all"). */
+export const PERIOD_HOUR_OPTIONS: readonly number[] = [24, 48, 72, 168];
