+++ package.json
@@ -9,7 +9,6 @@
     "test": "vitest run"
   },
   "dependencies": {
-    "cheerio": "^1.1.2",
     "luxon": "^3.5.0",
     "next": "14.2.13",
     "react": "18.3.1",
@@ -17,6 +16,7 @@
   },
   "devDependencies": {
     "@testing-library/react": "14.3.1",
+    "@types/luxon": "^3.4.2",
     "@types/node": "22.5.4",
     "@types/react": "18.3.3",
     "jsdom": "^27.0.0",
